[BUG] Fix delete operation
Use same Buffer if the page is already in buffer manager.
If loading same page more than once, there are mulitple pages
which share the same file contents. In this situation only page
which is flushed last is kept on the file and operations executed to
other pages are not saved.